refactor: decoupled tooltipHeaderFormatter selector
