+++ src/chart_types/xy_chart/state/chart_state.test.ts
@@ -8,14 +8,13 @@ import {
   SeriesTypes,
 } from '../utils/specs';
 import { Position } from '../../../utils/commons';
-import { TooltipType, TooltipValue } from '../utils/interactions';
 import { ScaleType, ScaleContinuous, ScaleBand } from '../../../scales';
 import { IndexedGeometry, GeometryValue, BandedAccessorType } from '../../../utils/geometry';
 import { AxisTicksDimensions, isDuplicateAxis } from '../utils/axis_utils';
 import { AxisId } from '../../../utils/ids';
 import { LegendItem } from '../legend/legend';
 import { ChartTypes } from '../..';
-import { SpecTypes } from '../../../specs/settings';
+import { SpecTypes, TooltipValue, TooltipType } from '../../../specs/settings';
 
 describe.skip('Chart Store', () => {
   let store: any = null; //
