+++ src/chart_types/xy_chart/state/selectors/is_tooltip_visible.ts
@@ -1,28 +1,21 @@
 import createCachedSelector from 're-reselect';
-import { TooltipType, isTooltipType, isTooltipProps } from '../../utils/interactions';
 import { Point } from '../../../../utils/point';
 import { GlobalChartState, PointerStates } from '../../../../state/chart_state';
 import { getSettingsSpecSelector } from '../../../../state/selectors/get_settings_specs';
 import { getProjectedPointerPositionSelector } from './get_projected_pointer_position';
 import { getTooltipValuesSelector, TooltipData } from './get_tooltip_values_highlighted_geoms';
 import { getChartIdSelector } from '../../../../state/selectors/get_chart_id';
+import { getTooltipType } from './get_tooltip_type';
+import { TooltipType } from '../../../../specs';
 
-const getTooltipType = (state: GlobalChartState): TooltipType | undefined => {
-  const tooltip = getSettingsSpecSelector(state).tooltip;
-  if (!tooltip) {
-    return undefined;
-  }
-  if (isTooltipType(tooltip)) {
-    return tooltip;
-  }
-  if (isTooltipProps(tooltip)) {
-    return tooltip.type;
-  }
+const hasTooltipTypeDefinedSelector = (state: GlobalChartState): TooltipType | undefined => {
+  return getTooltipType(getSettingsSpecSelector(state));
 };
+
 const getPointerSelector = (state: GlobalChartState) => state.interactions.pointer;
 
 export const isTooltipVisibleSelector = createCachedSelector(
-  [getTooltipType, getPointerSelector, getProjectedPointerPositionSelector, getTooltipValuesSelector],
+  [hasTooltipTypeDefinedSelector, getPointerSelector, getProjectedPointerPositionSelector, getTooltipValuesSelector],
   isTooltipVisible,
 )(getChartIdSelector);
