+++ src/chart_types/xy_chart/utils/interactions.ts
@@ -1,57 +1,6 @@
-import { $Values } from 'utility-types';
 import { Rotation } from '../../../utils/commons';
 import { Dimensions } from '../../../utils/dimensions';
 import { BarGeometry, PointGeometry, IndexedGeometry, isPointGeometry, isBarGeometry } from '../../../utils/geometry';
-import { TooltipProps } from '../../../specs';
-import { SeriesIdentifier } from './series';
-import { Accessor } from '../../../utils/accessor';
-
-/** The type of tooltip to use */
-export const TooltipType = Object.freeze({
-  /** Vertical cursor parallel to x axis */
-  VerticalCursor: 'vertical' as 'vertical',
-  /** Vertical and horizontal cursors */
-  Crosshairs: 'cross' as 'cross',
-  /** Follor the mouse coordinates */
-  Follow: 'follow' as 'follow',
-  /** Hide every tooltip */
-  None: 'none' as 'none',
-});
-
-export type TooltipType = $Values<typeof TooltipType>;
-
-export interface TooltipValue {
-  /**
-   * The label of the tooltip value
-   */
-  label: string;
-  /**
-   * The value to display
-   */
-  value: any;
-  /**
-   * The color of the graphic mark (by default the color of the series)
-   */
-  color: string;
-  /**
-   * True if the mouse is over the graphic mark connected to the tooltip
-   */
-  isHighlighted: boolean;
-  /**
-   * True if the tooltip is visible, false otherwise
-   */
-  isVisible: boolean;
-  /**
-   * The idenfitier of the related series
-   */
-  seriesIdentifier: SeriesIdentifier;
-  /**
-   * The accessor linked to the current tooltip value
-   */
-  valueAccessor: Accessor;
-}
-
-export type TooltipValueFormatter = (data: TooltipValue) => JSX.Element | string;
 
 /**
  * Get the cursor position depending on the chart rotation
@@ -83,13 +32,6 @@ export function getOrientedYPosition(xPos: number, yPos: number, chartRotation: 
     case 90:
       return chartDimension.width - xPos;
   }
-}
-
-export function isCrosshairTooltipType(type: TooltipType) {
-  return type === TooltipType.VerticalCursor || type === TooltipType.Crosshairs;
-}
-export function isFollowTooltipType(type: TooltipType) {
-  return type === TooltipType.Follow;
 }
 
 export function areIndexedGeometryArraysEquals(arr1: IndexedGeometry[], arr2: IndexedGeometry[]) {
@@ -132,12 +74,4 @@ function areBarEqual(ig1: BarGeometry, ig2: BarGeometry) {
     ig1.width === ig2.width &&
     ig1.height === ig2.height
   );
-}
-
-export function isTooltipProps(config: TooltipType | TooltipProps): config is TooltipProps {
-  return typeof config === 'object';
-}
-
-export function isTooltipType(config: TooltipType | TooltipProps): config is TooltipType {
-  return typeof config === 'string';
 }
