+++ src/components/tooltip/index.tsx
@@ -1,17 +1,17 @@
 import classNames from 'classnames';
 import React from 'react';
+import { createPortal } from 'react-dom';
 import { connect } from 'react-redux';
-import { TooltipValue, TooltipValueFormatter } from '../../chart_types/xy_chart/utils/interactions';
+import { TooltipValueFormatter, TooltipValue } from '../../specs';
 import { GlobalChartState, BackwardRef } from '../../state/chart_state';
+import { isInitialized } from '../../state/selectors/is_initialized';
+import { getTooltipHeaderFormatterSelector } from '../../state/selectors/get_tooltip_header_formatter';
 import { isTooltipVisibleSelector } from '../../chart_types/xy_chart/state/selectors/is_tooltip_visible';
-import { getTooltipHeaderFormatterSelector } from '../../chart_types/xy_chart/state/selectors/get_tooltip_header_formatter';
 import { getTooltipPositionSelector } from '../../chart_types/xy_chart/state/selectors/get_tooltip_position';
 import {
   getTooltipValuesSelector,
   TooltipData,
 } from '../../chart_types/xy_chart/state/selectors/get_tooltip_values_highlighted_geoms';
-import { isInitialized } from '../../state/selectors/is_initialized';
-import { createPortal } from 'react-dom';
 import { getFinalTooltipPosition, TooltipPosition } from '../../chart_types/xy_chart/crosshair/crosshair_utils';
 import { isAnnotationTooltipVisibleSelector } from '../../chart_types/xy_chart/state/selectors/is_annotation_tooltip_visible';
