+++ src/index.ts
@@ -16,7 +16,6 @@ export { SeriesCollectionValue } from './chart_types/xy_chart/utils/series';
 export { ChartTypes } from './chart_types';
 export { Datum, Position, Rendering, Rotation } from './utils/commons';
 export { TickFormatter } from './chart_types/xy_chart/utils/specs';
-export { TooltipType, TooltipValue, TooltipValueFormatter } from './chart_types/xy_chart/utils/interactions';
 export { SeriesIdentifier, XYChartSeriesIdentifier } from './chart_types/xy_chart/utils/series';
 export {
   AnnotationDomainType,
