+++ src/mocks/specs/specs.ts
@@ -13,8 +13,7 @@ import {
 import { getSpecId, getGroupId } from '../../utils/ids';
 import { ScaleType } from '../../scales';
 import { ChartTypes } from '../../chart_types';
-import { SettingsSpec, SpecTypes } from '../../specs';
-import { TooltipType } from '../../chart_types/xy_chart/utils/interactions';
+import { SettingsSpec, SpecTypes, TooltipType } from '../../specs';
 import { LIGHT_THEME } from '../../utils/themes/light_theme';
 
 export class MockSeriesSpec {
