+++ src/specs/settings.test.tsx
@@ -2,8 +2,7 @@ import { mount } from 'enzyme';
 import * as React from 'react';
 import { Position, Rendering, Rotation } from '../utils/commons';
 import { DARK_THEME } from '../utils/themes/dark_theme';
-import { TooltipType } from '../chart_types/xy_chart/utils/interactions';
-import { Settings, SettingsSpec } from './settings';
+import { Settings, SettingsSpec, TooltipType } from './settings';
 import { PartialTheme } from '../utils/themes/theme';
 import { LIGHT_THEME } from '../utils/themes/light_theme';
 import { chartStoreReducer, GlobalChartState } from '../state/chart_state';
