+++ src/specs/settings.tsx
@@ -2,15 +2,15 @@ import { $Values } from 'utility-types';
 import { DomainRange } from '../chart_types/xy_chart/utils/specs';
 import { PartialTheme, Theme } from '../utils/themes/theme';
 import { Domain } from '../utils/domain';
-import { TooltipType, TooltipValueFormatter } from '../chart_types/xy_chart/utils/interactions';
 import { getConnect, specComponentFactory } from '../state/spec_factory';
 import { Spec } from '.';
 import { LIGHT_THEME } from '../utils/themes/light_theme';
 import { ChartTypes } from '../chart_types';
 import { GeometryValue } from '../utils/geometry';
-import { XYChartSeriesIdentifier } from '../chart_types/xy_chart/utils/series';
+import { XYChartSeriesIdentifier, SeriesIdentifier } from '../chart_types/xy_chart/utils/series';
 import { Position, Rendering, Rotation } from '../utils/commons';
 import { ScaleContinuousType, ScaleOrdinalType } from '../scales';
+import { Accessor } from '../utils/accessor';
 
 export type ElementClickListener = (elements: Array<[GeometryValue, XYChartSeriesIdentifier]>) => void;
 export type ElementOverListener = (elements: Array<[GeometryValue, XYChartSeriesIdentifier]>) => void;
@@ -56,6 +56,53 @@ export interface PointerOutEvent extends BasePointerEvent {
 }
 
 export type PointerEvent = PointerOverEvent | PointerOutEvent;
+
+/** The type of tooltip to use */
+export const TooltipType = Object.freeze({
+  /** Vertical cursor parallel to x axis */
+  VerticalCursor: 'vertical' as 'vertical',
+  /** Vertical and horizontal cursors */
+  Crosshairs: 'cross' as 'cross',
+  /** Follor the mouse coordinates */
+  Follow: 'follow' as 'follow',
+  /** Hide every tooltip */
+  None: 'none' as 'none',
+});
+
+export type TooltipType = $Values<typeof TooltipType>;
+
+export interface TooltipValue {
+  /**
+   * The label of the tooltip value
+   */
+  label: string;
+  /**
+   * The value to display
+   */
+  value: any;
+  /**
+   * The color of the graphic mark (by default the color of the series)
+   */
+  color: string;
+  /**
+   * True if the mouse is over the graphic mark connected to the tooltip
+   */
+  isHighlighted: boolean;
+  /**
+   * True if the tooltip is visible, false otherwise
+   */
+  isVisible: boolean;
+  /**
+   * The idenfitier of the related series
+   */
+  seriesIdentifier: SeriesIdentifier;
+  /**
+   * The accessor linked to the current tooltip value
+   */
+  valueAccessor: Accessor;
+}
+
+export type TooltipValueFormatter = (data: TooltipValue) => JSX.Element | string;
 
 export interface TooltipProps {
   type?: TooltipType;
@@ -175,4 +222,20 @@ export function isPointerOutEvent(event: PointerEvent | null | undefined): event
 
 export function isPointerOverEvent(event: PointerEvent | null | undefined): event is PointerOverEvent {
   return event !== null && event !== undefined && event.type === PointerEventType.Over;
+}
+
+export function isTooltipProps(config: TooltipType | TooltipProps): config is TooltipProps {
+  return typeof config === 'object';
+}
+
+export function isTooltipType(config: TooltipType | TooltipProps): config is TooltipType {
+  return typeof config === 'string';
+}
+
+export function isCrosshairTooltipType(type: TooltipType) {
+  return type === TooltipType.VerticalCursor || type === TooltipType.Crosshairs;
+}
+
+export function isFollowTooltipType(type: TooltipType) {
+  return type === TooltipType.Follow;
 }
