+++ src/state/selectors/get_tooltip_header_formatter.ts
@@ -1,8 +1,7 @@
 import createCachedSelector from 're-reselect';
-import { isTooltipProps, TooltipValueFormatter } from '../../utils/interactions';
-import { getSettingsSpecSelector } from '../../../../state/selectors/get_settings_specs';
-import { SettingsSpec } from '../../../../specs/settings';
-import { getChartIdSelector } from '../../../../state/selectors/get_chart_id';
+import { getSettingsSpecSelector } from './get_settings_specs';
+import { SettingsSpec, TooltipValueFormatter, isTooltipProps } from '../../specs/settings';
+import { getChartIdSelector } from './get_chart_id';
 
 export const getTooltipHeaderFormatterSelector = createCachedSelector(
   [getSettingsSpecSelector],
